dom/tests/mochitest/cowl/test_external_confinement.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <p id="display"></p>
  <pre id="test">

  <script class="testbody" type="application/javascript">
  SimpleTest.waitForExplicitFinish();

  var completedTests = 0;
  var passedTests = 0;

  function ok_wrapper(result, desc) {
    ok(result, desc);

    completedTests++;

    if (result) {
      passedTests++;
    }

    if (completedTests == 4) {
      is(passedTests, completedTests, "There are " + completedTests + " external confinement tests that should pass");
      SimpleTest.finish();
    }
  }

  </script>
  <iframe src="file_external_confinement.html"></iframe>
  <iframe src="file_window_location_change.html"></iframe>
</pre>
</body>
</html>
